+++ bitcoin_nostr_chat/default_relays.py
@@ -4,17 +4,16 @@ from typing import List
 def get_preferred_relays() -> List[str]:
     return [
         "wss://nostr.btc-library.com",
-        "wss://nostr.hubmaker.io",
-        "wss://relay.primal.net",
         "wss://nostrue.com",
+        "wss://relay.primal.net",
         "wss://nostr.schneimi.de",
         "wss://nostr.sathoarder.com",
-        "wss://bitcoiner.social",
         "wss://nostr.swiss-enigma.ch",
-        "wss://nostr.blockpower.capital",
         "wss://ch.purplerelay.com",
         "wss://relay.poster.place",
-        "wss://nostr.satstralia.com",
+        "wss://nostr.middling.mydns.jp",
+        "wss://nostr.filmweb.pl",
+        "wss://nostr.openhoofd.nl",
     ]
 
 
@@ -23,23 +22,13 @@ def get_default_delays() -> List[str]:
         "wss://nostr.einundzwanzig.space",
         "wss://nostr.vulpem.com",
         "wss://nostr-1.nbo.angani.co",
-        "wss://relay.wellorder.net",
         "wss://relay.dwadziesciajeden.pl",
-        "wss://e.nos.lol",
         "wss://nostr.data.haus",
-        "wss://nostr.blockpower.capital",
-        "wss://nostr21.com",
         "wss://relay.nostromo.social",
         "wss://offchain.pub",
         "wss://nostrue.com",
         "wss://nproxy.kristapsk.lv",
-        "wss://nostr.topeth.info",
-        "wss://nostr.bitcoiner.social",
         "wss://nostr.roundrockbitcoiners.com",
-        "wss://nr.yay.so",
-        "wss://nostr.lu.ke",
-        "wss://bitcoiner.social",
-        "wss://relay.nostr.bg",
         "wss://slick.mjex.me",
         "wss://nostr.yael.at",
         "wss://relay.primal.net",
@@ -47,32 +36,21 @@ def get_default_delays() -> List[str]:
         "wss://nostr.swiss-enigma.ch",
         "wss://nostr.oxtr.dev",
         "wss://nostr.liberty.fans",
-        "wss://nostr.cro.social",
-        "wss://nostr.ginuerzh.xyz",
-        "wss://nostr.namek.link",
         "wss://nostr-02.dorafactory.org",
         "wss://relay.hodl.ar",
         "wss://nostr.middling.mydns.jp",
         "wss://nostr.filmweb.pl",
         "wss://nostr.openhoofd.nl",
-        "wss://rs2.abaiba.top",
         "wss://relay.poster.place",
-        "wss://ithurtswhenip.ee",
-        "wss://relay2.nostrchat.io",
-        "wss://relay1.nostrchat.io",
         "wss://relay.nostr.hu",
-        "wss://nostr.dakukitsune.ca",
         "wss://nostr.hifish.org",
         "wss://nostr2.sanhauf.com",
-        "wss://nostrua.com",
         "wss://relay.nostr.youlot.org",
-        "wss://relay.nostr.directory",
         "wss://relay.wavlake.com",
         "wss://nostr.fbxl.net",
         "wss://nostr.btc-library.com",
         "wss://relay.nostrcn.com",
         "wss://nostr-01.yakihonne.com",
-        "wss://relap.orzv.workers.dev",
         "wss://nostr.sathoarder.com",
         "wss://wbc.nostr1.com",
         "wss://nostr.heliodex.cf",
@@ -82,13 +60,9 @@ def get_default_delays() -> List[str]:
         "wss://nostr.dlsouza.lol",
         "wss://relay.kamp.site",
         "wss://relay.keychat.io",
-        "wss://nostr.2b9t.xyz",
-        "wss://relay.momostr.pink",
         "wss://unhostedwallet.com",
         "wss://21ideas.nostr1.com",
         "wss://nostr.stakey.net",
-        "wss://nostr.hubmaker.io",
-        "wss://unostr.site",
         "wss://frens.nostr1.com",
         "wss://nostr.dodge.me.uk",
         "wss://relay.magiccity.live",
@@ -97,17 +71,13 @@ def get_default_delays() -> List[str]:
         "wss://us.purplerelay.com",
         "wss://nostr1.daedaluslabs.io",
         "wss://freelay.sovbit.host",
-        "wss://nostr.huszonegy.world",
         "wss://dev-relay.kube.b-n.space",
         "wss://relay.nos.social",
-        "wss://ditto.slothy.win/relay",
         "wss://nostr.javi.space",
         "wss://gleasonator.dev/relay",
         "wss://bucket.coracle.social",
         "wss://relay.artx.market",
-        "wss://nfrelay.app",
         "wss://relay.notoshi.win",
-        "wss://relay.highlighter.com",
         "wss://nostr.petrkr.net/strfry",
         "wss://n.ok0.org",
         "wss://nostr-02.yakihonne.com",
@@ -118,51 +88,34 @@ def get_default_delays() -> List[str]:
         "wss://nostrich.adagio.tw",
         "wss://nostr.zoel.network",
         "wss://thecitadel.nostr1.com",
-        "wss://nostr2.daedaluslabs.io",
         "wss://fiatjaf.nostr1.com",
-        "wss://nostr.daedaluslabs.io",
         "wss://nostr.gerbils.online",
-        "wss://jingle.carlos-cdb.top",
-        "wss://carlos-cdb.top",
         "wss://jp.purplerelay.com",
         "wss://ir.purplerelay.com",
         "wss://me.purplerelay.com",
         "wss://nostr.0x7e.xyz",
         "wss://strfry.openhoofd.nl",
-        "wss://nostr.kolbers.de",
         "wss://relay.nostr.net",
         "wss://v1250.planz.io/nostr",
         "wss://relay.13room.space",
         "wss://relay.usefusion.ai",
         "wss://ae.purplerelay.com",
-        "wss://nostrasia.mom",
         "wss://au.purplerelay.com",
         "wss://in.purplerelay.com",
-        "wss://nostr.satstralia.com",
         "wss://bostr.bitcointxoko.com",
-        "wss://test.nfrelay.app",
         "wss://relay.fanfares.io",
         "wss://relay.varke.eu",
         "wss://relay.nostr.jabber.ch",
         "wss://ftp.halifax.rwth-aachen.de/nostr",
         "wss://tw.purplerelay.com",
         "wss://eu.purplerelay.com",
-        "wss://testnet.plebnet.dev/nostrrelay/1",
         "wss://hivetalk.nostr1.com",
-        "wss://relay.earthly.land",
-        "wss://relay.piazza.today",
-        "wss://bostr.nokotaro.work",
         "wss://orangepiller.org",
         "wss://ch.purplerelay.com",
-        "wss://bostr.nokotaro.com",
-        "wss://nostr.animeomake.com",
-        "wss://nostr.openordex.org",
         "wss://lnbits.papersats.io/nostrclient/api/v1/relay",
         "wss://strfry.chatbett.de",
         "wss://nostr.plantroon.com",
         "wss://relay.corpum.com",
-        "wss://relay.proxymana.net",
-        "wss://nsrelay.assilvestrar.club",
         "wss://nostr.atitlan.io",
         "wss://relay.livefreebtc.dev",
         "wss://nostr.tavux.tech",
@@ -170,7 +123,6 @@ def get_default_delays() -> List[str]:
         "wss://relay.nostrich.cc",
         "wss://support.nostr1.com",
         "wss://ca.purplerelay.com",
-        "wss://chefstr.nostr1.com",
         "wss://nostr.babyshark.win",
         "wss://relay.agorist.space",
         "wss://relay.nostrhub.fr",
@@ -182,37 +134,29 @@ def get_default_delays() -> List[str]:
         "wss://bostr.lightningspore.com",
         "wss://de.purplerelay.com",
         "wss://obiurgator.thewhall.com",
-        "wss://slime.church/relay",
         "wss://relay.unknown.cloud",
         "wss://relay.gems.xyz",
         "wss://onlynotes.lol",
-        "wss://relay2.nostrasia.net",
         "wss://relay.zone667.com",
         "wss://relay.mostro.network",
         "wss://relay5.bitransfer.org",
-        "wss://brisceaux.com",
         "wss://nostr.polonkai.eu",
         "wss://libretechsystems.nostr1.com",
-        "wss://nostr.kloudcover.com",
         "wss://relay.alex71btc.com",
-        "wss://nostr3.daedaluslabs.io",
         "wss://relay1.xfire.to:",
         "wss://nostr.brackrat.com",
         "wss://relay.satlantis.io",
         "wss://relay.illuminodes.com",
         "wss://relay.arrakis.lat",
-        "wss://nostr.manasiwibi.com",
         "wss://junxingwang.org",
         "wss://relay2.angor.io",
         "wss://history.nostr.watch",
         "wss://relay.angor.io",
         "wss://nostr.tbai.me:592",
-        "wss://strfry.iris.to",
         "wss://orangesync.tech",
         "wss://nostr.chaima.info",
         "wss://relay.minibolt.info",
         "wss://kr.purplerelay.com",
-        "wss://relay.tapestry.ninja",
         "wss://fl.purplerelay.com",
         "wss://nostr.bilthon.dev",
         "wss://relay.gasteazi.net",
@@ -228,12 +172,9 @@ def get_default_delays() -> List[str]:
         "wss://relay.lnfi.network",
         "wss://relay.nostrdice.com",
         "wss://nostr.azzamo.net",
-        "wss://wheat.happytavern.co",
         "wss://strfry.bonsai.com",
-        "wss://devs.nostr1.com",
         "wss://relay.notestack.com",
         "wss://relay.sigit.io",
-        "wss://nostr.noderunners.network",
         "wss://thebarn.nostr1.com",
         "wss://nostr.grooveix.com",
         "wss://travis-shears-nostr-relay-v2.fly.dev",
@@ -242,25 +183,18 @@ def get_default_delays() -> List[str]:
         "wss://art.nostrfreaks.com",
         "wss://cobrafuma.com/relay",
         "wss://alru07.nostr1.com",
-        "wss://relay.flirtingwithbitcoin.com",
-        "wss://relay.nostrpunk.com",
         "wss://relay.botev.sv",
         "wss://relay.degmods.com",
-        "wss://untreu.me",
         "wss://reimagine.nostr1.com",
         "wss://nostr.takasaki.dev",
-        "wss://nostr.coincrowd.fund",
-        "wss://nostr.inaridiy.com",
         "wss://nostr.community.ath.cx",
         "wss://relay.xeble.me",
         "wss://strfry.shock.network",
         "wss://relay02.lnfi.network",
-        "wss://filter.weme.wtf",
         "wss://relay.nostrfreaks.com",
         "wss://cyberspace.nostr1.com",
         "wss://nostr-rs-relay.dev.fedibtc.com",
         "wss://relay.cyphernomad.com",
-        "wss://relay.jellyfish.land",
         "wss://articles.layer3.news",
         "wss://ghost.dolu.dev",
         "wss://thebarn.nostrfreaks.com",
@@ -269,59 +203,35 @@ def get_default_delays() -> List[str]:
         "wss://nostr.mikoshi.de",
         "wss://relay.nuts.cash",
         "wss://nostr.agentcampfire.com",
-        "wss://nostr.me/relay",
         "wss://dwebcamp.nos.social",
         "wss://henhouse.social/relay",
         "wss://relay.stewlab.win",
         "wss://relay.crbl.io",
         "wss://rl.baud.one",
-        "wss://schnorr.me",
         "wss://relay.axeldolce.xyz",
         "wss://relay.i9.eti.br",
-        "wss://immo.jellyfish.land",
         "wss://nostr.schneimi.de",
-        "wss://relay.hunstr.mywire.org",
         "wss://pareto.nostr1.com",
         "wss://bonifatius.nostr1.com",
         "wss://relay.xplbzx.uk",
         "wss://nostr.tac.lol",
-        "wss://btcpay2.nisaba.solutions/nostr",
-        "wss://nostream-stg.ordlify.com",
         "wss://dev-relay.lnfi.network",
-        "wss://nostr.allgemeinheit.info",
-        "wss://straylight.cafe/relay",
         "wss://relay.bitcoinveneto.org",
-        "wss://relay.gnostr.cloud",
-        "wss://nostr-relay.apps.den.ratwater.xyz",
-        "wss://nostr.red5d.dev",
         "wss://relay-testnet.k8s.layer3.news",
-        "wss://nostr.pistaum.com",
         "wss://relay-nwc.rizful.com/v1",
         "wss://eupo43gj24.execute-api.us-east-1.amazonaws.com/test",
         "wss://backup.keychat.io",
-        "wss://relay.newatlantis.top",
         "wss://relay.openbalance.app",
-        "wss://nostr-relay.kloudcover.com",
         "wss://social.proxymana.net",
         "wss://inbox.azzamo.net",
         "wss://nostr.lojong.info",
-        "wss://nostr.nwe.li",
         "wss://nostr-rs-relay-ishosta.phamthanh.me",
         "wss://nostr.rblb.it:7777",
-        "wss://nostr.basedpotato.org",
         "wss://sources.nostr1.com",
         "wss://relay.nostronautti.fi",
-        "wss://nostr.d3id.xyz/relay",
         "wss://relay.nostrtalk.org",
-        "wss://ditto.openbalance.app/relay",
-        "wss://relay.tv-base.com",
-        "wss://nostr.itdestro.cc",
-        "wss://nostr.tegila.com.br",
-        "wss://nostream.oviato.com",
-        "wss://relay.devs.tools",
         "wss://relay.stream.labs.h3.se",
         "wss://mls.akdeniz.edu.tr/nostr",
         "wss://tollbooth.stens.dev",
-        "wss://nostr.casaalmida.com",
         "wss://nostrich.zonemix.tech",
     ]
