+++ bitcoin_nostr_chat/signals_min.py
@@ -29,9 +29,9 @@
 
 import logging
 
-logger = logging.getLogger(__name__)
-
 from PyQt6.QtCore import QObject, QThread, pyqtSignal
+
+logger = logging.getLogger(__name__)
 
 
 class SignalsMin(QObject):
