+++ pyproject.toml
@@ -11,7 +11,7 @@ disable_error_code = "assignment"
 
 [tool.poetry]
 name = "bitcoin-nostr-chat"
-version = "0.5.1"
+version = "0.6.0"
 authors = ["andreasgriffin <andreasgriffin@proton.me>"]
 license = "GPL-3.0"
 readme = "README.md"
@@ -22,9 +22,9 @@ homepage = "https://github.com/andreasgriffin/bitcoin-nostr-chat"
 python = ">=3.9,<3.13"
 cbor2 = "^5.6.3"
 requests = "^2.31.0"
-pyqt6 = "^6.6.1"
-nostr-sdk = "^0.32.1"
+pyqt6 = "^6.6.1" 
 bitcoin-qr-tools = "^1.0.0"
+nostr-sdk = "^0.38.0"
