+++ tests/test_relays.py
@@ -5,11 +5,11 @@ from typing import List
 import bdkpython as bdk
 import pytest
 from nostr_sdk import Keys
-from PyQt6.QtCore import QObject, pyqtSignal
+from PyQt6.QtCore import QCoreApplication, QObject, pyqtSignal
 from pytestqt.qtbot import QtBot
 
 from bitcoin_nostr_chat.default_relays import get_default_delays, get_preferred_relays
-from bitcoin_nostr_chat.nostr import BitcoinDM, ChatLabel, DmConnection, RelayList
+from bitcoin_nostr_chat.group_chat import BitcoinDM, ChatLabel, DmConnection, RelayList
 
 logger = logging.getLogger(__name__)
 
@@ -27,14 +27,14 @@ logger = logging.getLogger(__name__)
 #         pytest.fail(f"WebSocket connection failed for {url}: {str(e)}")
 
 
-class TestClass(QObject):
+class DummyClass(QObject):
     signal_dm = pyqtSignal(BitcoinDM)
 
 
-def send_dms(qtbot: QtBot, relays: List[str], raise_error: bool):
+def send_dms_to_self(qtbot: QtBot, relays: List[str], raise_error: bool):
     network = bdk.Network.REGTEST
 
-    test_instance = TestClass()
+    test_instance_recipient = DummyClass()
     from_serialized = lambda base64_encoded_data: BitcoinDM.from_serialized(
         base64_encoded_data, network=network
     )
@@ -49,15 +49,14 @@ def send_dms(qtbot: QtBot, relays: List[str], raise_error: bool):
         logger.info(f"relay: {relay}")
 
         dm_connection = DmConnection(
-            test_instance.signal_dm,
+            test_instance_recipient.signal_dm,
             from_serialized=from_serialized,
             keys=keys,
             get_currently_allowed=get_currently_allowed,
             relay_list=RelayList(relays=[relay], last_updated=datetime.now(), max_age=5000),
         )
-        dm_connections.append(dm_connection)
         dm_connection.subscribe()
-
+        dm_connections.append(dm_connection)
         dm = BitcoinDM(
             label=ChatLabel.GroupChat,
             created_at=datetime.now(),
@@ -68,30 +67,39 @@ def send_dms(qtbot: QtBot, relays: List[str], raise_error: bool):
             logger.info(f"Success: {dm.description}")
             successful_relays.append(relay)
 
-        test_instance.signal_dm.connect(add_to_chat)
+        test_instance_recipient.signal_dm.connect(add_to_chat)
 
         try:
-            with qtbot.waitSignal(test_instance.signal_dm, timeout=10000):
+            with qtbot.waitSignal(test_instance_recipient.signal_dm, timeout=10000):
                 dm_connection.send(dm, receiver=keys.public_key())
         except Exception as e:
             logger.error(f"not working: {relay}")
             if raise_error:
                 raise Exception(f"not working: {relay} , original_message {e}")
         finally:
-            dm_connection.disconnect()
+
+            QCoreApplication.processEvents()  # Allow Qt to process events
+            # sender.stop()
             dm_connection.stop()
-            test_instance.signal_dm.disconnect(add_to_chat)
+            test_instance_recipient.signal_dm.disconnect(add_to_chat)
+            QCoreApplication.processEvents()  # Allow Qt to process events
 
     logger.info(f"=" * 50)
     logger.info(f"Good relays: {successful_relays}")
     logger.info(f"=" * 50)
 
 
+# def test_send_dms_umbrel(
+#     qtbot: QtBot,
+# ) -> None:
+#     send_dms_to_self(qtbot, relays=["ws://umbrel:4848"], raise_error=True)
+
+
 def test_send_dms_preffered(
     qtbot: QtBot,
 ) -> None:
 
-    send_dms(qtbot, relays=get_preferred_relays(), raise_error=True)
+    send_dms_to_self(qtbot, relays=get_preferred_relays(), raise_error=True)
 
 
 @pytest.mark.testrelays
@@ -99,4 +107,4 @@ def test_send_dms(
     qtbot: QtBot,
 ) -> None:
 
-    send_dms(qtbot, relays=get_default_delays(), raise_error=False)
+    send_dms_to_self(qtbot, relays=get_default_delays(), raise_error=False)
